requirements.txt
streamlit
langchain-google-genai
langchain-community
langchain
chromadb
pypdf
google-generativeai